Roll back ovs-p4rt refactoring (#490)
* Revert "ovsp4rt housekeeping (10/10) (#477)"

This reverts commit 892b1ea8bb9d44f9f45b3d3aaafa1a600d6c50a7.

* Revert "Extract public common functions (9/10) (#476)"

This reverts commit 1861f03c0bb68cbd5f3a5c82196222ee79520d61.

* Revert "Extract public ES2K functions (#475)"

This reverts commit b6db3a4cedf53184e862d7d44ba8e20c129ecc52.

* Revert "Extract private ES2K functions (#473)"

This reverts commit 5e2e07e7fb42bf7e510491de56fda417b87e8e33.

* Revert "Extract public DPDK functions (#472)"

This reverts commit 7d305d30482fe317cfa9d5e0b025b9e52b004021.

* Revert "Extract private DPDK functions (#471)"

This reverts commit 3c2c8379b4bfe313b6d5ef4ade9247ff6cae333b.

* Revert "Publish common function prototypes (#470)"

This reverts commit 0d08c880025a7c73a2f748d1314e3a63c61df541.

* Revert "ovsp4rt housekeeping (#469)"

This reverts commit 6763e06d6b7fa0d2b4819d00420a9d1fb286ddac.

* Revert "Move utility functions to ovsp4rt_utils (#468)"

This reverts commit db090afeb0a27cb8b756ec36b925cdbeef7c30bb.

* Revert "Move common ovsp4rt files to 'common' subdirectory (#467)"

This reverts commit b74d83d231529840fb2659f4cab66ea9b50cf353.

* Revert ovs submodule reference to 623d2da6557ef2236bc04ee99a1097a0f940380e


---------
