+++ ovs-p4rt/sidecar/CMakeLists.txt
@@ -4,26 +4,28 @@
 # SPDX-License-Identifier: Apache 2.0
 #
 
+#################
+# ovs_sidecar_o #
+#################
+
 add_library(ovs_sidecar_o OBJECT
-    ${OVSP4RT_INCLUDE_DIR}/ovsp4rt/ovs-p4rt.h
+    ovsp4rt.cc
+    ovsp4rt_credentials.cc
+    ovsp4rt_credentials.h
+    ovsp4rt_session.cc
+    ovsp4rt_session.h
 )
-
-target_include_directories(ovs_sidecar_o
-    PUBLIC
-        ${OVSP4RT_INCLUDE_DIR}
-    PRIVATE
-        ${CMAKE_CURRENT_SOURCE_DIR}
-        ${STRATUM_SOURCE_DIR}
-)
-
-add_subdirectory(common)
-add_subdirectory(lib)
 
 if(DPDK_TARGET)
     add_subdirectory(dpdk)
 elseif(ES2K_TARGET)
     add_subdirectory(es2k)
 endif()
+
+target_include_directories(ovs_sidecar_o PUBLIC
+    ${STRATUM_SOURCE_DIR}
+    ${OVS_INCLUDE_DIR}
+)
 
 target_link_libraries(ovs_sidecar_o PUBLIC
     absl::strings
@@ -49,17 +51,4 @@ target_link_libraries(ovs_sidecar PUBLIC
 
 add_dependencies(ovs_sidecar ovs_sidecar_o)
 
-install(
-    DIRECTORY
-        ${OVSP4RT_INCLUDE_DIR}/ovsp4rt
-    DESTINATION
-        ${CMAKE_INSTALL_INCLUDEDIR}
-)
-
-###########
-# testing #
-###########
-
-set(SIDECAR_SOURCE_DIR ${CMAKE_CURRENT_SOURCE_DIR})
-
-add_subdirectory(testing EXCLUDE_FROM_ALL)
+set_install_rpath(ovs_sidecar $ORIGIN ${SDE_ELEMENT} ${DEP_ELEMENT})
