+++ ovs-p4rt/sidecar/es2k/CMakeLists.txt
@@ -4,8 +4,5 @@
 # SPDX-License-Identifier: Apache 2.0
 
 target_sources(ovs_sidecar_o PRIVATE
-    ovsp4rt_es2k_defs.h
-    ovsp4rt_es2k_private.cc
-    ovsp4rt_es2k_private.h
-    ovsp4rt_es2k_public.cc
+    p4_name_mapping.h
 )
