+++ ovs-p4rt/sidecar/ovsp4rt_session.h
@@ -36,7 +36,7 @@ class OvsP4rtSession {
  public:
   // Create the session with given P4runtime stub and device id
   static ::absl::StatusOr<std::unique_ptr<OvsP4rtSession>> Create(
-      std::unique_ptr<::p4::v1::P4Runtime::Stub> stub, uint32_t device_id,
+      std::unique_ptr<p4::v1::P4Runtime::Stub> stub, uint32_t device_id,
       const std::string& role_name,
       ::absl::uint128 election_id = TimeBasedElectionId());
 
@@ -60,13 +60,13 @@ class OvsP4rtSession {
 
   std::string RoleName() const { return role_name_; }
 
-  ::p4::v1::Uint128 ElectionId() const { return election_id_; }
+  p4::v1::Uint128 ElectionId() const { return election_id_; }
 
-  ::p4::v1::P4Runtime::Stub& Stub() { return *stub_; }
+  p4::v1::P4Runtime::Stub& Stub() { return *stub_; }
 
  private:
   OvsP4rtSession(uint32_t device_id, std::string role_name,
-                 std::unique_ptr<::p4::v1::P4Runtime::Stub> stub,
+                 std::unique_ptr<p4::v1::P4Runtime::Stub> stub,
                  ::absl::uint128 election_id)
       : device_id_(device_id),
         stub_(std::move(stub)),
@@ -78,33 +78,33 @@ class OvsP4rtSession {
 
   uint32_t device_id_;
 
-  ::p4::v1::Uint128 election_id_;
+  p4::v1::Uint128 election_id_;
 
   std::string role_name_;
 
-  std::unique_ptr<::p4::v1::P4Runtime::Stub> stub_;
+  std::unique_ptr<p4::v1::P4Runtime::Stub> stub_;
 
   std::unique_ptr<grpc::ClientContext> stream_channel_context_;
 
-  std::unique_ptr<grpc::ClientReaderWriter<::p4::v1::StreamMessageRequest,
-                                           ::p4::v1::StreamMessageResponse>>
+  std::unique_ptr<grpc::ClientReaderWriter<p4::v1::StreamMessageRequest,
+                                           p4::v1::StreamMessageResponse>>
       stream_channel_;
 };
 
-std::unique_ptr<::p4::v1::P4Runtime::Stub> CreateP4RuntimeStub(
+std::unique_ptr<p4::v1::P4Runtime::Stub> CreateP4RuntimeStub(
     const std::string& address,
     const std::shared_ptr<grpc::ChannelCredentials>& credentials);
 
 // Functions that operate on a OvsP4rtSession.
 
-::absl::StatusOr<::p4::v1::ReadResponse> SendReadRequest(
-    OvsP4rtSession* session, const ::p4::v1::ReadRequest& read_request);
+::absl::StatusOr<p4::v1::ReadResponse> SendReadRequest(
+    OvsP4rtSession* session, const p4::v1::ReadRequest& read_request);
 
 ::absl::Status SendWriteRequest(OvsP4rtSession* session,
-                                const ::p4::v1::WriteRequest& write_request);
+                                const p4::v1::WriteRequest& write_request);
 
 ::absl::Status GetForwardingPipelineConfig(OvsP4rtSession* session,
-                                           ::p4::config::v1::P4Info* p4info);
+                                           p4::config::v1::P4Info* p4info);
 
 ::p4::v1::TableEntry* SetupTableEntryToInsert(OvsP4rtSession* session,
                                               ::p4::v1::WriteRequest* req);
